% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_elements.R
\name{common.elements}
\alias{common.elements}
\title{Common elements}
\usage{
common.elements(a, b)
}
\arguments{
\item{a}{first vector}

\item{b}{second vector}
}
\description{
Finds if vectors have common elements.
}

