% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deform_elements.R
\name{deform.elements}
\alias{deform.elements}
\title{Deform object divided into elements}
\usage{
deform.elements(data, elements.list, a, theta, random.sd)
}
\arguments{
\item{data}{data matrix}

\item{elements.list}{list of elements (landmark numbers)}

\item{a}{deformation amount}

\item{theta}{deformation angle}

\item{random.sd}{deformation random sd}
}
\description{
Deforms object with randomized strain matrix, randomization is common for landmarks in the same element.
}

