% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deform_population.R
\name{deform.population}
\alias{deform.population}
\title{Deform population}
\usage{
deform.population(data, min.a = 0.8, max.a = 1.2, min.theta = -0.5,
  max.theta = 0.5, diff.elements = FALSE, curves = NULL,
  split.list = NULL, random.sd = NULL)
}
\arguments{
\item{data}{data array}

\item{min.a}{minimal deformation amount; default = 0.8}

\item{max.a}{maximum deformation amount; default = 1.2}

\item{min.theta}{minimal deformation angle; default = -0.5}

\item{max.theta}{maximum deformation angle; default = 0.5}

\item{diff.elements}{should elements of single object have slightly diffrent, random deformation; default = FALSE}

\item{curves}{optional curves matrix for creatin elements list; default = NULL}

\item{split.list}{opional list of elements; default = NULL}

\item{random.sd}{optional standard deviation of randomness in element version; default = NULL}
}
\description{
Deforms population in random way.
}

