% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_elements.R
\name{find.elements}
\alias{find.elements}
\title{Search for elements}
\usage{
find.elements(curves)
}
\arguments{
\item{curves}{curves matrix}
}
\description{
Search for elements in semilandmark matrix.
}

