% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_model_response_from_coeff.R
\name{find.model.response.from.coeff}
\alias{find.model.response.from.coeff}
\title{Solves equation}
\usage{
find.model.response.from.coeff(interception, slope, x = NA, y = NA)
}
\arguments{
\item{interception}{interception}

\item{slope}{slope}

\item{x}{x}

\item{y}{y}
}
\description{
Solves equation y = ax + b
}

