% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model.R
\name{generate.model}
\alias{generate.model}
\title{Create linear model}
\usage{
generate.model(data)
}
\arguments{
\item{data}{landmarks array}
}
\description{
Creates linear of landmarks with size as predictor
}

