% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_predicted.R
\name{generate.predicted}
\alias{generate.predicted}
\title{Create artificial landmark set}
\usage{
generate.predicted(model, predictor, random.sd = 1e-04, elements = FALSE,
  elements.list = NULL)
}
\arguments{
\item{model}{model}

\item{predictor}{predictor}

\item{random.sd}{standard deviation of coefficients randomness; default = 0.0001}

\item{elements}{randomization constant in elements; default = FALSE}

\item{elements.list}{optional list of elements; default = NULL}
}
\description{
Generates artificial landmark set based on model.
}

