% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_population.R
\name{sample.population}
\alias{sample.population}
\title{Create an artificial population}
\usage{
sample.population(model, population.size = 10, Csize = c(10, 100),
  curves = NULL, symmetry = FALSE, mid = NULL, left = NULL,
  right = NULL, random.sd = 1e-04, elements = FALSE,
  elements.list = NULL)
}
\arguments{
\item{model}{model}

\item{population.size}{size of population}

\item{Csize}{two element vector containing minimum and maximum of centroid size}

\item{curves}{optional curves matrix}

\item{symmetry}{optional symmetrization of landmarks sets}

\item{mid}{midline for symmerizaion}

\item{left}{leftside landmarks for symmetrization}

\item{right}{rightside landmarks for symmerization}

\item{random.sd}{standard deviation for randomization of model}

\item{elements}{optional randomization constant in elements}

\item{elements.list}{list of elements}
}
\description{
Creates an artificial population of landmark sets from model
}

