% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_list.R
\name{split.list}
\alias{split.list}
\title{Split landmarks matrix into elements}
\usage{
\method{split}{list}(data, curves)
}
\arguments{
\item{data}{matrix of landmarks}

\item{curves}{curves matrix}
}
\description{
Splits landmark matrix into list of elements
}

